+++ src/app/contracts/resilienceWebProxy.ts
@@ -1,5 +1,6 @@
 import axios = require("axios");
 import { IMaintance } from "./maintenance";
+import { Guid } from "guid-typescript";
 
 /**
  * A proxy for a cached web resiliency operation.
@@ -9,9 +10,10 @@ export interface IResilienceWebProxy {
      * Executes a function within a web resilience proxy. Uses optionally a cache.
      * @param request Web request to send.
      * @param cacheKey Key for the cache if enabled.
+     * @param guid Request Guid.
      * @returns The result of the executed function.
      */
-    execute<TResult>(request: axios.AxiosRequestConfig, cacheKey?: string): Promise<axios.AxiosResponse<TResult>>;
+    execute<TResult>(request: axios.AxiosRequestConfig, cacheKey?: string, guid?: Guid): Promise<axios.AxiosResponse<TResult>>;
 
     /**
      * Gets the maintenance.